Module Parallelmodel_lib.Ast

type expr =

| INT of int

| VAR_NAME of string	(*
variable shared by multiple threads
*)

| REGISTER of string	(*
thread local registers
*)

| PLUS of expr * expr

| SUB of expr * expr

| MUL of expr * expr

| DIV of expr * expr

val pp_expr : Ppx_show_runtime.Format.formatter -> expr -> unit

val show_expr : expr -> string

type stmt =

| ASSIGN of expr * expr

| IF of expr * stmt list	(*
if expr <> 0 then execute a list of statements within curly brackets
*)

| IF_ELSE of expr * stmt list * stmt list	(*
if expr <> 0 then execute the first list of statements within curly brackets else execute the second one
*)

| SMP_MB	(*
memory barrier that flushes store buffer
*)

| NO_OP

val pp_stmt : Ppx_show_runtime.Format.formatter -> stmt -> unit

val show_stmt : stmt -> string

type thread =

| THREAD of int * stmt list

val pp_thread : Ppx_show_runtime.Format.formatter -> thread -> unit

val show_thread : thread -> string

type prog =

| PROG of thread list

val pp_prog : Ppx_show_runtime.Format.formatter -> prog -> unit

val show_prog : prog -> string