yocaml_cmarkit

API

Library yocaml_cmarkit

Yocaml_cmarkit Describing documents using a Markup language is very common in the Blogosphere (rather than describing all the formatting of document content in HTML, using <p>, <strong> and co).